+++ src/bpp/exceptions.py
@@ -1,0 +1,2 @@
+class UczelniaNotDefined(Exception):
+    pass